+++ src/App.js
@@ -11,6 +11,7 @@ import { getRegistry } from '@redhat-cloud-services/frontend-components-utilitie
 import { notificationsReducer } from '@redhat-cloud-services/frontend-components-notifications/redux';
 import { Bullseye, Spinner } from '@patternfly/react-core';
 import LockIcon from '@patternfly/react-icons/dist/js/icons/lock-icon';
+import useChrome from '@redhat-cloud-services/frontend-components/useChrome';
 
 import { Routes } from './Routes';
 import ErrorBoundary from './Utilities/ErrorBoundary';
@@ -23,26 +24,31 @@ const App = ({ useLogger, basename }) => {
   const history = useHistory();
   const [isAuthenticated, setIsAuthenticated] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
+  const chrome = useChrome();
 
   useEffect(() => {
-    const registry = getRegistry();
-    registry.register({ notifications: notificationsReducer });
-    insights.chrome.init();
-    insights.chrome.auth.getUser().then(() => {
-      setIsAuthenticated(true);
-      setIsLoading(false);
-    });
-    insights.chrome.identifyApp('ocp-advisor');
-    const unregister = insights.chrome.on('APP_NAVIGATION', (event) => {
-      const targetUrl = event.domEvent?.href
-        ?.replace(basename, '/')
-        .replace(/^\/\//, '/');
-      if (typeof targetUrl === 'string') {
-        history.push(targetUrl);
-      }
-    });
+    let unregister;
+    if (chrome) {
+      const registry = getRegistry();
+      registry.register({ notifications: notificationsReducer });
+      const { on: onChromeEvent } = chrome.init();
+
+      unregister = onChromeEvent('APP_NAVIGATION', (event) => {
+        const targetUrl = event.domEvent?.href
+          ?.replace(basename, '/')
+          .replace(/^\/\//, '/');
+        if (typeof targetUrl === 'string') {
+          history.push(targetUrl);
+        }
+      });
+      chrome.auth.getUser().then(() => {
+        setIsAuthenticated(true);
+        setIsLoading(false);
+      });
+    }
     return () => unregister();
-  }, []);
+  }, [chrome]);
+
   return (
     <ErrorBoundary>
       {isLoading ? (