+++ src/Components/Recommendation/Recommendation.js
@@ -15,7 +15,6 @@ import {
   PageHeader,
   PageHeaderTitle,
 } from '@redhat-cloud-services/frontend-components/PageHeader';
-import { Main } from '@redhat-cloud-services/frontend-components/Main';
 import { DateFormat } from '@redhat-cloud-services/frontend-components/DateFormat';
 import {
   Label,
@@ -40,7 +39,11 @@ import {
 
 import Breadcrumbs from '../Breadcrumbs';
 import RuleLabels from '../Labels/RuleLabels';
-import { FILTER_CATEGORIES, RULE_CATEGORIES } from '../../AppConstants';
+import {
+  FILTER_CATEGORIES,
+  RISK_OF_CHANGE_DESC,
+  RULE_CATEGORIES,
+} from '../../AppConstants';
 import messages from '../../Messages';
 import Loading from '../Loading/Loading';
 import { adjustOCPRule } from '../../Utilities/Rule';
@@ -53,6 +56,7 @@ import ViewHostAcks from '../Modals/ViewHostAcks';
 import { OneLineLoader } from '../../Utilities/Loaders';
 import { enableRuleForCluster } from '../../Services/Acks';
 import { formatMessages, mapContentToValues } from '../../Utilities/intlHelper';
+import inRange from 'lodash/inRange';
 
 const Recommendation = ({ rule, ack, clusters, match }) => {
   const intl = useIntl();
@@ -174,18 +178,18 @@ const Recommendation = ({ rule, ack, clusters, match }) => {
         <Breadcrumbs current={content?.description || recId} />
       </PageHeader>
       {(isUninitialized || isLoading || isFetching) && (
-        <Main>
+        <section className="pf-l-page__main-section pf-c-page__main-section pf-m-light pf-u-pt-sm">
           <Loading />
-        </Main>
+        </section>
       )}
       {isError && (
-        <Main>
+        <section className="pf-l-page__main-section pf-c-page__main-section pf-m-light pf-u-pt-sm">
           <ErrorState />
-        </Main>
+        </section>
       )}
       {!(isUninitialized || isLoading || isFetching) && isSuccess && (
         <React.Fragment>
-          <Main className="pf-m-light pf-u-pt-sm">
+          <section className="pf-l-page__main-section pf-c-page__main-section pf-m-light pf-u-pt-sm">
             <RuleDetails
               messages={formatMessages(
                 intl,
@@ -248,6 +252,13 @@ const Recommendation = ({ rule, ack, clusters, match }) => {
               onVoteClick={async (rule, rating) =>
                 await Post(`${BASE_URL}/v2/rating`, {}, { rule, rating })
               }
+              {...(inRange(content?.resolution_risk, 1, 5) // resolution risk can be 0 (not defined for particular rule)
+                ? {
+                    resolutionRisk: content?.resolution_risk,
+                    resolutionRiskDesc:
+                      RISK_OF_CHANGE_DESC[content?.resolution_risk],
+                  }
+                : {})}
             >
               <Flex>
                 <FlexItem align={{ default: 'alignRight' }}>
@@ -298,8 +309,8 @@ const Recommendation = ({ rule, ack, clusters, match }) => {
                 </FlexItem>
               </Flex>
             </RuleDetails>
-          </Main>
-          <Main>
+          </section>
+          <section className="pf-l-page__main-section pf-c-page__main-section">
             <React.Fragment>
               {(content?.hosts_acked_count ||
                 ackedClusters?.length > 0 ||
@@ -434,7 +445,7 @@ const Recommendation = ({ rule, ack, clusters, match }) => {
                 />
               )}
             </React.Fragment>
-          </Main>
+          </section>
         </React.Fragment>
       )}
     </React.Fragment>