+++ src/Components/RecsList/index.js
@@ -1,12 +1,12 @@
 import React, { lazy, Suspense } from 'react';
 import { useIntl } from 'react-intl';
 
-import Main from '@redhat-cloud-services/frontend-components/Main';
 import PageHeader from '@redhat-cloud-services/frontend-components/PageHeader';
 
 import Loading from '../Loading/Loading';
 import messages from '../../Messages';
 import { Title } from '@patternfly/react-core';
+import useChrome from '@redhat-cloud-services/frontend-components/useChrome';
 
 const RecsListTable = lazy(() =>
   import(/* webpackChunkName: 'RulesTable' */ '../RecsListTable/')
@@ -14,7 +14,9 @@ const RecsListTable = lazy(() =>
 
 const RecsList = () => {
   const intl = useIntl();
-  insights.chrome.updateDocumentTitle(
+  const chrome = useChrome();
+
+  chrome.updateDocumentTitle(
     intl.formatMessage(messages.documentTitle, {
       subnav: intl.formatMessage(messages.recommendations),
     })
@@ -29,11 +31,11 @@ const RecsList = () => {
             .toLowerCase()}`}
         </Title>
       </PageHeader>
-      <Main>
+      <section className="pf-l-page__main-section pf-c-page__main-section">
         <Suspense fallback={<Loading />}>
           <RecsListTable />
         </Suspense>
-      </Main>
+      </section>
     </React.Fragment>
   );
 };