+++ src/Components/RecsListTable/RecsListTable.cy.js
@@ -8,6 +8,7 @@ import { RecsListTable } from './RecsListTable';
 import getStore from '../../Store';
 import ruleResponse from '../../../cypress/fixtures/api/insights-results-aggregator/v2/rule.json';
 import { Intl } from '../../Utilities/intlHelper';
+// TODO: use cypress utils from FEC
 import {
   TOOLBAR,
   TOOLBAR_FILTER,
@@ -73,6 +74,7 @@ const data = _.orderBy(values, ['total_risk'], ['desc']);
 
 const IMPACT = { Low: 1, Medium: 2, High: 3, Critical: 4 };
 const LIKELIHOOD = { Low: 1, Medium: 2, High: 3, Critical: 4 };
+const RISK_OF_CHANGE = { 'Very Low': 1, Low: 2, Moderate: 3, High: 4 };
 const STATUS = ['All', 'Enabled', 'Disabled'];
 const IMPACTING = { '1 or more': 'true', None: 'false' };
 const CATEGORIES_MAP = {
@@ -120,6 +122,16 @@ const filtersConf = {
     urlParam: 'likelihood',
     urlValue: (it) =>
       encodeURIComponent(_.map(it, (x) => LIKELIHOOD[x]).join(',')),
+  },
+  res_risk: {
+    selectorText: 'Risk of change',
+    values: Array.from(cumulativeCombinations(Object.keys(RISK_OF_CHANGE))),
+    type: 'checkbox',
+    filterFunc: (it, value) =>
+      _.map(value, (x) => RISK_OF_CHANGE[x]).includes(it.resolution_risk),
+    urlParam: 'res_risk',
+    urlValue: (it) =>
+      encodeURIComponent(_.map(it, (x) => RISK_OF_CHANGE[x]).join(',')),
   },
   category: {
     selectorText: 'Category',
@@ -242,6 +254,23 @@ describe('data', () => {
     expect(
       _.filter(filterData(), (it) => it.impacted_clusters_count > 1)
     ).to.have.length.gte(1);
+  });
+  it('at least one recommendation in default list has resolution risk set to non 0 value', () => {
+    expect(
+      filterData().filter(
+        ({ resolution_risk, description }) =>
+          resolution_risk >= 1 &&
+          description.includes('1Lorem ipsum dolor sit amet')
+      ).length
+    ).to.gte(1);
+  });
+  it('at least one recommendation in default list has resolution risk set to 0', () => {
+    expect(
+      filterData().filter(
+        ({ resolution_risk, description }) =>
+          resolution_risk === 0 && description.includes('Super atomic nuclear')
+      ).length
+    ).to.gte(1);
   });
 });
 
@@ -471,6 +500,7 @@ describe('successful non-empty recommendations list table', () => {
         'publish_date',
         'tags',
         'total_risk',
+        'resolution_risk',
         'impacted_clusters_count',
       ],
       TABLE_HEADERS
@@ -488,6 +518,7 @@ describe('successful non-empty recommendations list table', () => {
                 )
               );
           }
+
           checkSorting(
             filterData(DEFAULT_FILTERS, dataUnsorted),
             sortingParameter,
@@ -496,7 +527,7 @@ describe('successful non-empty recommendations list table', () => {
             'Name',
             'description',
             DEFAULT_DISPLAYED_SIZE,
-            label
+            label === 'Risk of change' ? 'res_risk' : label
           );
         });
       });
@@ -638,6 +669,42 @@ describe('successful non-empty recommendations list table', () => {
       cy.get('th[data-label="Name"]')
         .should('have.attr', 'aria-sort')
         .and('contain', 'ascending');
+    });
+
+    it('absent resolution risk is shown as not available', () => {
+      cy.getRowByName(
+        'Super atomic nuclear cluster on the brink of the world destruction'
+      )
+        .find('[data-label="Risk of change"]')
+        .should('have.text', 'Not available');
+    });
+
+    it('absent resolution risk is not shown in expanded details', () => {
+      cy.getRowByName(
+        'Super atomic nuclear cluster on the brink of the world destruction'
+      )
+        .find('.pf-c-table__toggle button')
+        .click();
+      cy.get(EXPANDABLES)
+        .eq(0)
+        .find('.pf-m-spacer-sm')
+        .contains('Risk of change')
+        .should('not.exist');
+    });
+
+    it('present resolution risk is shown in expanded details', () => {
+      cy.getRowByName('1Lorem ipsum dolor sit amet')
+        .find('.pf-c-table__toggle button')
+        .click();
+      cy.get(EXPANDABLES)
+        .eq(0)
+        .find('.pf-m-spacer-sm')
+        .contains('Risk of change')
+        .should('exist');
+      cy.get(EXPANDABLES)
+        .eq(0)
+        .find('.ins-c-rule-details__risk-of-ch-label')
+        .should('have.text', 'High');
     });
   });